+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/change/tracking/spi/display/LayoutCTDisplayRenderer.java
@@ -46,7 +46,7 @@ import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.segments.constants.SegmentsWebKeys;
 import com.liferay.segments.service.SegmentsExperienceLocalService;
 import com.liferay.style.book.model.StyleBookEntry;
-import com.liferay.style.book.util.DefaultStyleBookEntryUtil;
+import com.liferay.style.book.util.StyleBookEntryProviderUtil;
 
 import jakarta.portlet.PortletRequest;
 
@@ -247,7 +247,7 @@ public class LayoutCTDisplayRenderer extends BaseCTDisplayRenderer<Layout> {
 			"style-book",
 			() -> {
 				StyleBookEntry styleBookEntry =
-					DefaultStyleBookEntryUtil.getStyleBookEntry(layout);
+					StyleBookEntryProviderUtil.getStyleBookEntry(layout);
 
 				if (styleBookEntry == null) {
 					return null;
